Updated the logic around our Skybrud link picker models a bit
- Added "LinkPickerItem.Udi" property
The Umbraco 8 package saves an UDI instead of a numeric ID as the Umbraco 7 package did. To still support Umbraco 7 link picker models, the "Id" property is still available.

- "LinkPickerItem.Url" property shouldn't be nullable
In the Umbraco 8 link picker item model, the URL may be left out if an UDI is available instead.

- "LinkPickerItem.Target" property shouldn't be nullable
The target may the omitted.

- Renamed "LinkPickerItem.Mode" property to "LinkPickerItem.Type"
For the Umbraco 8 package we renamed the property as well as the corresponding enum. The Umbraco 7 "mode" value is still supported, and will be used as fallback if a "type" property isn't available.

- etc.