Introduced new "LegacyObjectBase" class to help with nullability decorations
Most models derived from an instance of "JObject" would either extend the "JsonObjectBase" class from our Skybrud.Essentials package, or no extend any classes at all. The "JsonObjectBase" class exposed the underlying "JObject" instance, which is what we wan't, but given the implementation of the "JsonObjectBase" class, the "JObject" property may be null.

But with the way that we're using it, we know that the "JObject" is never null, and as such, the "LegacyObjectBase" now has a "JObject" property that is not null.