Updated the SkybrudGridDataParser to be more in control of the initializing returned models
Previously the invidual models would have the parser as a parameter, and then handle some of the parsing within their constructors. This is generally considered bad practice - e.g. because it makes it harder to overwrite the logic. With this commit, that logic is now moved to the parser instead.